Refactor(data): Rename weight field to be more specific and add mappers
Renamed the `weight` field in `ExerciseEntity` to `weightInKg` to clarify that the weight is stored in kilograms. This change is reflected in the database schema and test data.

Additionally, this commit introduces mapper functions to convert data transfer objects (DTOs) to domain models. These mappers handle the transformation of entities like `ExerciseDto`, `RoutineDto`, and `RoutineItemDto` into their corresponding domain-layer representations, including a new `Weight` value class.